audio-api: Fix linting errors after adding `sbt-tpolecat`

+++ audio-api/src/main/scala/no/ndla/audioapi/service/search/AudioSearchService.scala
@@ -50,7 +50,7 @@ trait AudioSearchService {
 
       val fullSearch = settings.query match {
         case Some(query) =>
-          val languageSearch = (field: String, boost: Float) =>
+          val languageSearch = (field: String, boost: Double) =>
             languageSpecificSearch(
               field,
               settings.language,
